"Decent" fix for particle duration bug
We're experiencing another bug with libgdx,  particles (dynamic
animations) are rendering way too long, we're now manually checking &
forcing effects to be done instead of using the libraries
implementation. After inspecting the source of libgdx, we noticed that
we can get the effect we desire by checking if an animation has gone
from -100% to 0% done, rather than the default -100% -> infinity. This
may cause some effects in some properties of particle-effects.