ControlPanel is now BoardObject, updated BoardObjec
The ControlPanel now extends BoardObject. The BoardObject is updated so
it can have a separate width and height that does not depend on the
image for the BoardObject. I have also added another contructor for
BoardObject that can take width and height as parameters as well. If no
width and height is entered these are taken from the Image.