Added functionality for selecting and placing towers
Added the Hand (for holding objects) and the Holdable interface that
holdable objects should implement. Made Tower into an Holdable object.
HandController was created for updating the position of the hand when
the mouse is moved. Hand was put into GameData so it is reachable from
everywhere. Also made BoardObject and Tower implement the
"cloneable"-interface. When the control-panel is clicked the tower of
the ControlPanelTower is cloned and put in the hand. The hand can then
place the Tower on the map. Image also implements cloneable.