Restructuring to prepare for Spell Panel
Restructured how the ControlPanels work to make it easier to put in more
panels. Panels are now handled in class ControlPanelController instead
of GameScreen and are placed in a list for easier access from the
Renderer-class. Added two constants to Constants class in the process
and also renamed the filePath-constant to FILE_PATH.