Refactoring: Moving functionality to Map
Moving data to Map to more correctly follow the principle "Tell don't
ask". It is no longer possible to get lists of projectiles and such from
the map to for example render or update instead the map is told to
render, or told to update. Remove and add-stack is now handled in Map.