Spell cooldowns
This took embarrassing amounts of effort on my part, this solution is
somewhat dirty since every concrete spell implements its own cooldown
system (which does allow lots of flexebility, but causes duplicate
code). Putting the cooldown system in the abstract class spell causes a
lot of errors, depending on how you do it:

Setting a static value of cooldown in spell (an int) would cause all
spells to share the cooldown

Creating a cooldownManager item would require a new
cooldownmanagerController that decreases the spell cooldowns OR making
the cooldownmanager a persistentitem, but it has no position or image,
and the methods would still have to be written in the concrete spell
classes.

Putting the cooldowns in the ControlPanelSpell reduces flexebility, and
would create complex code situations since how it would be started and
managed is not clear.